requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv
requests
google-genai==1.51.0
